% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_2021_2023}
\alias{aus_2021_2023}
\title{Life Tables of Australian States}
\format{
\subsection{\code{aus_2021_2023}}{

A data frame with 1818 rows and 12 columns
\describe{
\item{Stat}{Indigenous status. \code{Aboriginal} or \code{Non-Aboriginal}}
\item{Sex}{Sex. \code{Female} or \code{Male}}
\item{Age}{Age in single years form 0 to 100. The last age group is open.}
\item{mx}{Death rate}
\item{qx}{Death probability}
\item{ax}{Separation factor. The average number of additional years lived by those who died at this age}
\item{lx}{Survival fraction}
\item{dx}{Deaths}
\item{Lx}{Person years lived}
\item{Tx}{Person years lived after}
\item{ex}{Life expectancy}
\item{OpenInterval}{Logical. Flogs whether an interval is open (\code{TRUE}) or closed (\code{FALSE})}
}
}
}
\usage{
aus_2021_2023
}
\description{
Life tables of Australian states by sex.
}
\keyword{datasets}
