% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedisp.R
\name{lifedisp}
\alias{lifedisp}
\title{Adds a life disparity column to a life table}
\usage{
lifedisp(lt)
}
\arguments{
\item{lt}{A life table}
}
\value{
A life table with the extra column \code{vx} giving the life disparity at age \code{x}
}
\description{
Adds a life disparity column to a life table
}
