% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ogompertz}
\alias{ogompertz}
\title{Calibrate Gompertz model to last two age intervals of a life table}
\usage{
ogompertz(q, n, olifeexp)
}
\arguments{
\item{q}{Death probability for the last closed interval}

\item{n}{Age interval for the last closed interval}

\item{olifeexp}{Life expectancy at the start of the open interval}
}
\value{
A list of Gompertz parameters
}
\description{
Calibrate Gompertz model to last two age intervals of a life table
}
\examples{
ogompertz(0.3,5,10)
}
