% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedisp.R
\name{olifedisp}
\alias{olifedisp}
\title{Returns life disparity at the open interval using Gompertz model}
\usage{
olifedisp(q, n, olifeexp)
}
\arguments{
\item{q}{Death probability for the last closed interval}

\item{n}{n Age interval for the last closed interval}

\item{olifeexp}{olifeexp Life expectancy at the start of the open interval}
}
\value{
life disparity
}
\description{
Returns life disparity at the open interval using Gompertz model
}
\examples{
olifedisp(0.3,5,10)
}
