% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedisp.R
\name{vfunc_sqa}
\alias{vfunc_sqa}
\title{Returns normalised V from the A and B parameters of the SQA approximation}
\usage{
vfunc_sqa(A, B)
}
\arguments{
\item{A}{Linear coefficient of the survival fraction}

\item{B}{Quadratic coefficient of the survival fraction}
}
\value{
Normalised V. V = Normalised V x AgeInterval * SurvFrac
}
\description{
Returns normalised V from the A and B parameters of the SQA approximation
}
