% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedisp.R
\name{lifedisp}
\alias{lifedisp}
\title{Adds a life disparity column to a life table}
\usage{
lifedisp(lt, method = "sqa")
}
\arguments{
\item{lt}{A life table}

\item{method}{The approximatiom methed. Can be \code{'sqa'}, \code{'enu'}, \code{'els'} or \code{'els'}. Defaults to \code{'sqa'}.}
}
\value{
A life table with the extra column \code{vx} giving the life disparity at age \code{x}
}
\description{
Adds a life disparity column to a life table
}
\examples{
library(dplyr)
daus_lt <- aus_2021_2023 \%>\% group_by(State,Sex) \%>\% lifedisp()
}
